requirements.txt
# Web Server
flask
gunicorn

# Data Handling
pandas
openpyxl
xlrd

# Plotting & Graphics
matplotlib

# Utilities
requests
semver

# PostgreSQL Connector
psycopg2-binary

selenium

Flask-Login

Flask-Bcrypt

alembic

python-dotenv

SQLAlchemy

